Optimize session storage for users w/o cookies
This cuts down drastically on the amount of data saved to Redis and the
traffic to and from Redis by avoiding querying for a session that we
know won't be there, such as on the first time a user visits the site.

Previously, we created a session if it didn't exist, which would create
sessions for bots and other users who would never visit the site again,
wasting a whole lot of Redis storage and generating a lot of unnecessary
traffic to the Redis server and making the app wait for the results of
those session queries. Now we only persist the session if it has been
modified, such as when requesting a form which generates a CSRF token.