Optimize form_params slightly
Previously it would run a regex matcher, when all we needed was to check
for inclusion of a substring. Regexes are slower than substring matches.

This commit also moves this from manual memoization into a getter with a
block for memoization. This isn't an optimization for performance as
much as an optimization for reading.