Fix occasional compile-time failures
The Crystal compiler sometimes has issues with downcasting the
`IO::Memory` into an `IO` because that's actually an upcast, not a
downcast. It doesn't happen all the time, though, and I don't know what
triggers it, but I did notice that this fixes it in the cases that do
trigger that failure.