Allow static route matches with slashes
Example:
    # matches `GET /foo/bar/baz/quux`
    r.on "foo/bar/baz/quux" do
      # ...
    end

This commit also optimizes path matching to be about 33x faster. Not 33%
faster, but 33 *times*. It's not that regexes in Crystal are slow, it's
just that not using them is *SO MUCH FASTER*. Processing a request in
6µs is pretty decent, but doing it in 188ns is even better.

Before:

```
nested 161.71k (  6.18µs) (± 1.64%)  3.44kB/op  fastest
```

After:

```
nested   5.31M (188.18ns) (± 1.18%)  464B/op  fastest
```

And when we match without nesting, we can optimize it even further:

```
static  15.85M ( 63.11ns) (± 0.96%)  240B/op        fastest
```

That is 98x as fast as the nested matchers prior to this commit.